Watch WIP
Can now safely start/stop watch and declare which files to watch or not
watch on a test by test basis.

Next step is to make the monitor function loop through the watchList and
run everything in it if one of the files has been modified.

Then I'll just need to add the ability to add a specific watchFile or
watchFolder.