watchFile, watchFolder, VERBOSE
Now with watchFile and watchFolder, it's easy to turn the monitor_ctrl
on and leave the tests up while working in the code. You get emediate
feedback if something you change breaks a test.

Also added a verbose feature that is off be default. Turn it on to see a
full list of the files watched.