WIP Branching.js
Finally got things reorganized so that I can test Branching. Pulled out
the file moc from log.test into it's own testing module File.moc.js

Changed Branching to take in a 'logger' which for now will just be
Log.js which allows me to enstantiate it with the mocFile in the test.
This also opens the door for different 'loggers' to be written for the
branching module. So as long as the logger in question supports the same
api someone could write a logger that interfaces with a webAPI etc...