Delay.js
Ok so Timers/setTimeout and setInterval don't really make sense because
I'm only able to impliment $.sleep in Illustrator and the real set-
timeoutand setInterval are supposed to be Asynchronus.

So in order to avoid confusion later on, I changed the names to Delay,
doAfter and repeatAfter. This way they are clearly something seperate
and hopefully less lickly to be miss-interpreted as asynchronus.

To be clear, this is BLOCKING code, not good but it's all we have for
the watch functionality.

100% code coverate