WIP watcher.js
Only had a little bit of time today, set up the final test and almost
finished the logic. This module needs to require the delay module and
responsibilaty of quitting the tests needs to be moved to the main file.
For now I've commented out app.quite and left a 'FIXME' reminder.