Trying an NPM approach
After a couple weeks of being too busy/deliberating as to how this
framework should finally be peaced together. I've come to the conclusion
that an NPM CLI is the way to go. So this is the first commit in a series
of steps towards wireing all the tested modules up into one awesome CLI
for testing. Best part, now I can use the ready-made ai-deftly tests as
fodder whilest peacing this thing together!