ParseArguments and ShowHelp
Added parseArguments.sj to bin and modified index.js to use it.
This function process the arguments array and attaches them to an object.
Currently two types of arguments are recognized:
  Flags
  - the default type with a value of true
  KeyValue
  - denoted by two dashes at the beginning of the e.g. --key value

More types can easily be added by establishing conventions for the first
two characters in a name. e.g. v3Myvector3 0 0 0 or v2MyVector2 0 0

The help.js file is simply for keeping things neat and organized.
It contains the help text for testing-deftly and supplies a show()
method for console logging it out to the user.