WIP Can now get options from command line and .opts file
Getting the .opts file should be possible by simply using require. Try
again to get the path and pass it to require. The error I was having
before was simply because the json had an extra comma at the end.

Also look into getting the calling directory (pwd) without using an
exec command